changes admin.py files
