+++ django_36/book/admin.py
@@ -1,3 +1,8 @@
 from django.contrib import admin
+from book.models import Books
 
-# Register your models here.
+
+@admin.register(Books)
+class BooksAdmin(admin.ModelAdmin):
+    list_display = ('id', 'title', 'author', 'year', 'price')
+    search_fields = ('title', 'author')
