+++ django_36/purchase/admin.py
@@ -1,3 +1,8 @@
 from django.contrib import admin
+from purchase.models import Purchases
 
-# Register your models here.
+
+@admin.register(Purchases)
+class PurchasesAdmin(admin.ModelAdmin):
+    list_display = ('id', 'user', 'book', 'date')
+    search_fields = ('user__first_name', 'user__last_name', 'book__title')
